Implement the mocked function testcases.
The simnowdata provides the expected raw_cpuid response for a given
output for each type of Cpu.
The MockX86CpuTestSuite test the UserApis with the data from the simnow
folder against the provided expected output.
CpuidUtilsTestSuite Implementation(imcomplete to be completed in the
following patch) tests the CupidUtils functions in a similar fashion.