Add hasFlags testcase
Two testcases hasFlagsT and hasFlagsF added for both qemu and mock
testsuites.

New .txt files FlagsT.txt and FlagsF.txt are added for all the cpuTypes
in simnowdata folder. All the Flags in FlagsT.txt is expected to be
present in that particular cpuType. All flags in FlagsF.txt is expected
to be absent in that particular cpuType.