Add QemuTestcase to verify the vendorInfo returned from the capi
Moved the QemuTest Class to be the baseclass for already existing qemu
tests,
Added new class for vendorInfo test.

The test calls the au_cpuid_get_vendor api and writes the value to a
file for every emulated cpu.
Verifies it against the expected vendorInfo values in the Qemutestcase.

Some clean up.